feat: refactor toast message handling and move related functions to utils
